services/dispenser/requirements.txt
-i https://pypi.org/simple
cachetools==2.1.0
google-api-core==1.3.0
grpcio==1.14.2
mysql-connector==2.2.9
protobuf==3.6.1
pymongo==3.9.0
transitions==0.6.8
rq==1.1.0
